requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai>=0.8.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
python-telegram-bot>=20.0
